// crm_frontend_3/src/api/api.ts
export interface Order_Client_DetailsVm {
    iD_Order: number;
    name_Client: string;
    lastName_Client: string;
    email_Client?: string | undefined;
    telefon: string;
    type_technology: string;
    model_technology: string;
    breaking_info: string;
    quipment_info: string ;
    status_Order?: string | undefined;
    iD_Personnel_dispatcher: string;
    iD_Personnel_master: string;
    receipt_date: Date;
    issue_date?: Date | undefined;
    price?: number;
}

export interface List_Personnal {
    iD_Personnal: string;
    name: string;
    last_Name: string;
    telefon: string;
    position: string;
}

export interface List_PersonnalVm {
    list_Personnals: List_Personnal[] ;
}


export interface PersonnalVm {
    name: string ;
    last_Name: string ;
    email?: string | undefined;
    telefon: string ;
    position?: string | undefined;
}

export interface ProblemDetails {
    type?: string | undefined;
    title?: string | undefined;
    status?: number | undefined;
    detail?: string | undefined;
    instance?: string | undefined;
}

export interface Update_Order_ClientDto {
    iD_Order:number;
    status_Order: string ;
}

export interface Update_Status_orderDto {
    iD_Order: number;
   
}