// crm_frontend_3/src/api/api3.ts
export interface List_Equipment {
    iD_Order: number;
    iD_Accessories: number;
    name_Сomponent: string ;
    type_Сomponent: string ;
    price_Сomponent: number;
    name_Company: string;
}